pyproject.toml
[project]
name = "project-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime==4.11",
    "dotenv>=0.9.9",
    "google-api-core>=2.25.0",
    "google-generativeai>=0.8.5",
    "googleapis-common-protos>=1.70.0",
    "nougat-ocr>=0.1.17",
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pypdf2>=3.0.1",
    "sympy>=1.14.0",
]
